Make range count operator save bounds when count is 0
The range count operator previously always returned the range 1..0 if the
count is 0.  It would be preferable to save specified bounds so that e.g.
an empty range could easily be declared with a given low bound (3..#0).

Now, when the count is 0 for the range '#' operator, if the range is
boundedLow, the result range is low..low-1.  If the range is boundedHigh,
the result is high+1..high.  If high+1 or low-1 are out of bounds for the
type, it is a runtime error (halt). If the range is fully bounded, the
low bound is saved when the stride is positive and the high bound is
saved if the stride is negative.

Removed a .future asking for this change.