Move chpl__defaultHash for records from compiler generated to the modules
The chpl__defaultHash function for records can be implemented in module code
using functions from the Reflection module.  Remove the code that generates
it from the compiler and add it to the DefaultAssociative module.

I tested that this returns the same value as the compiler generated version
for records containing 0, 1, 2, 3 and 4 ints, reals, and complexes.