Fix a bug involving returning the void value from functions.
A function that returned '_void' from any return was made to only return
void. Remove the code that did that to allow returning '_void' or not in
different branches of a param conditional.

Once that was done, the .future fixed in this commit ran into an issue in
postFold because '_void' was not considered a param. Update the function
issuing the error to allow '_void' as a param.

While here, cleaned these functions up a bit.