Avoid assigning calls to functions that return no value to variables
Since the return value goes to a retTemp even when it's void now, a few spots
in the compiler didn't recognize it as void and assigned calls to the function
to a temp.

Key off of FLAG_VOID_NO_RETURN_VALUE to signal that these are void and take
the void code path.