Differentiate between void values and empty returns
Void values and functions that don't return anything are two different
concepts, so give them distinct types.  Void values can be assigned to
variables, which causes the compiler to remove the variable.  A function that
doesn't return any value cannot be assigned to a variable.

A function that returns nothing needs to return nothing on all code paths.
A function that returns a void value can return a different type on a different
code path as long as there is a param controlling which path is chosen.

Most of the code for handling void values was updated to work with gVoidVal and
dtVoidVal instead of gVoid and dtVoid.