Checkpoint an unfinished JSON formatter
This is a start to a JSON formatter for the standard IO. It implements a
handful of the functions in the interface and tries using them to do IO for
default associative arrays.

At the moment it is failing to work because the formatter field in the
channel is becoming nil by the time the it gets to the `dsiSerialReadWrite()`
method.