Allow unbounded and partially bounded ranges for array declarations
Add a transformation for array declarations to the cleanup pass that passes
the result of the initialization expression to chpl__ensureDomainExpr.
chpl__ensureDomainExpr can then use the initialization to determine the size
of the array, and create a bounded range to use as the actual array type.

Issue an error for an array with an unbounded or partially bounded domain that
is not explicitly initialized.

Add tests that declare arrays with bounded low/high/unbounded specified
domains. Add a test that trips the new error for no initializer.