Fix some unbounded arrays issues
The wrappers for handling unbounded array declarations were causing problems
with multidimensional arrays looking like associative arrays. Rework them
and rename them to better handle this.

Moving the init expression to a temp before the array def caused problems
with arrays declared in classes or records. Change it to instead be the init
expression for the temp.