Bug fixes for moving lvalueCheck
Change an assert to be true when run outside of function resolution.

Move the whole lvalueCheck to the start of lateConstCheck, instead of doing
it in-line with the rest of the checks.

This still has around 40 failures due to less specific error messages than
before.

Signed off by: David Iten daviditen@users.noreply.github.com